+++ R/task_spec.R
@@ -15,7 +15,7 @@ task_spec <- function(alias = NULL, package_spec = NULL, env = NULL) {
     list(
       alias = alias,
       package_spec = package_spec,
-      env = env
+      env = merge_default_configuration(env, DEFAULT_R_CMD_CHECK_VARIABLES)
     ),
     class = "task_spec"
   )
@@ -84,8 +84,8 @@ custom_install_task_spec <- function(...) {
 check_task_spec <- function(args = NULL, build_args = NULL, ...) {
   task_spec <- task_spec(...)
   check_spec <- list(
-    args = args,
-    build_args = build_args
+    args = merge_default_configuration(args, DEFAULT_CHECK_ARGS),
+    build_args = merge_default_configuration(build_args, DEFAULT_BUILD_ARGS)
   )
 
   structure(